+++ ha1/src/main/java/Square.java
@@ -11,31 +11,31 @@ public class Square {
 
     private Point topLeft;
 
-    public double getSideLength() {
+    public float getSideLength() {
         return sideLength;
     }
 
-    public void setSideLength(double sideLength) {
+    public void setSideLength(float sideLength) {
         this.sideLength = sideLength;
     }
 
-    private double sideLength;
+    private float sideLength;
 
     public Square() {
         this.topLeft = new Point(0,0);
         this.sideLength = 1;
     }
 
-    public Square(Point topLeft, double side) {
+    public Square(Point topLeft, float side) {
         this.topLeft = topLeft;
         this.sideLength = side;
     }
 
-    public static Square fromArea(Point topLeft, double area) {
-        return new Square(topLeft, Math.sqrt(area));
+    public static Square fromArea(Point topLeft, float area) {
+        return new Square(topLeft, (float)Math.sqrt(area));
     }
 
-    public double calculateArea() {
+    public float calculateArea() {
         return this.sideLength * this.sideLength;
     }
